refactor: unify openai parsers under modelship.openai.parsers
Move modelship/openai/tool_calling/ to modelship/openai/parsers/tool_calling/
to make room for a sibling reasoning/ package. Extracts read_chat_template
(and its GGUF + tokenizer_config helpers) to a shared parsers/utils.py since
both tool-call and reasoning parser detection key off the chat template.

No behavior change.
