+++ modelship/openai/parsers/tool_calling/__init__.py
@@ -7,10 +7,10 @@ already emit structured tool calls (vLLM, llama.cpp via a function-calling
 chat handler) bypass it.
 """
 
-from modelship.openai.tool_calling.input import resolve_tools_for_request
-from modelship.openai.tool_calling.parsers import ParsedToolCalls, ToolCallParser, ToolCallStreamer
-from modelship.openai.tool_calling.registry import available_parsers, get_parser, register_parser
-from modelship.openai.tool_calling.streaming import (
+from modelship.openai.parsers.tool_calling.input import resolve_tools_for_request
+from modelship.openai.parsers.tool_calling.parsers import ParsedToolCalls, ToolCallParser, ToolCallStreamer
+from modelship.openai.parsers.tool_calling.registry import available_parsers, get_parser, register_parser
+from modelship.openai.parsers.tool_calling.streaming import (
     build_chat_completion_response,
     finish_reason_for,
     stream_chat_completion,
